+++ tests/testthat/test-utils_etc.R
@@ -12,7 +12,7 @@ p <- 10L
 m <- 100L
 
 test_that("Random z have right output dim and the sums are between 1 and p-1", {
-  Z <- sample_Z(p, m = m)
+  Z <- sample_Z(p, m = m, feature_names = LETTERS[1:p])
   
   expect_equal(dim(Z), c(m, p))
   expect_true(all(rowSums(Z) %in% 1:(p - 1L)))
@@ -20,23 +20,27 @@ test_that("Random z have right output dim and the sums are between 1 and p-1", {
 
 test_that("Random z have right output dim and the sums are in subset S", {
   S <- 2:3
-  Z <- sample_Z(p, m = m, S = S)
+  Z <- sample_Z(p, m = m, feature_names = LETTERS[1:p], S = S)
   
   expect_equal(dim(Z), c(m, p))
   expect_true(all(rowSums(Z) %in% S))
 })
 
 test_that("Sampling input structure is ok (deg = 0)", {
-  input <- input_sampling(p, m = m, deg = 0L, paired = TRUE)
+  input <- input_sampling(
+    p, m = m, deg = 0L, paired = TRUE, feature_names = LETTERS[1:p]
+  )
 
   expect_equal(dim(input$Z), c(m, p))
   expect_equal(sum(input$w), 1.0)
   expect_equal(dim(input$A), c(p, p))
-  expect_equal(diag(input$A), rep(0.5, p))
+  expect_equal(unname(diag(input$A)), rep(0.5, p))
 })
 
 test_that("Sampling input structure is ok (deg = 0, unpaired)", {
-  input <- input_sampling(p, m = m, deg = 0L, paired = FALSE)
+  input <- input_sampling(
+    p, m = m, deg = 0L, paired = FALSE, feature_names = LETTERS[1:p]
+  )
   
   expect_equal(dim(input$Z), c(m, p))
   expect_equal(sum(input$w), 1.0)
@@ -45,7 +49,9 @@ test_that("Sampling input structure is ok (deg = 0, unpaired)", {
 })
 
 test_that("Sampling input structure is ok (deg = 1)", {
-  input <- input_sampling(p, m = m, deg = 1L, paired = TRUE)
+  input <- input_sampling(
+    p, m = m, deg = 1L, paired = TRUE, feature_names = LETTERS[1:p]
+  )
   
   expect_equal(dim(input$Z), c(m, p))
   expect_true(sum(input$w) < 1.0)
@@ -54,7 +60,9 @@ test_that("Sampling input structure is ok (deg = 1)", {
 })
 
 test_that("Sampling input input structure ok (deg = 2)", {
-  input <- input_sampling(p, m = m, deg = 2L, paired = TRUE)
+  input <- input_sampling(
+    p, m = m, deg = 2L, paired = TRUE, feature_names = LETTERS[1:p]
+  )
   
   expect_equal(dim(input$Z), c(m, p))
   expect_true(sum(input$w) < 1.0)
@@ -64,8 +72,8 @@ test_that("Sampling input input structure ok (deg = 2)", {
 
 test_that("Partly exact A, w, Z equal exact for sufficiently large deg", {
   for (p in 2:10) {
-    pa <- input_partly_exact(p, deg = trunc(p / 2))
-    ex <- input_exact(p)
+    pa <- input_partly_exact(p, deg = trunc(p / 2), feature_names = LETTERS[1:p])
+    ex <- input_exact(p, feature_names = LETTERS[1:p])
     pa_rs <- rowSums(pa$Z)
     ex_rs <- rowSums(ex$Z)
     
@@ -77,39 +85,42 @@ test_that("Partly exact A, w, Z equal exact for sufficiently large deg", {
 
 test_that("hybrid weights sum to 1 for different p and degree 1", {
   deg <- 1L
-  expect_error(input_sampling(2L, deg = deg))
-  expect_error(input_sampling(3L, deg = deg))
+  expect_error(input_sampling(2L, deg = deg, feature_names = LETTERS[1:p]))
+  expect_error(input_sampling(3L, deg = deg, feature_names = LETTERS[1:p]))
 
   for (p in 4:20) {
-    pa <- input_partly_exact(p, deg = deg)
-    sa <- input_sampling(p, m = 100L, deg = deg, paired = TRUE)
+    pa <- input_partly_exact(p, deg = deg, feature_names = LETTERS[1:p])
+    sa <- input_sampling(
+      p, m = 100L, deg = deg, paired = TRUE, feature_names = LETTERS[1:p]
+    )
     expect_equal(sum(pa$w) + sum(sa$w), 1.0)
   }
 })
 
 test_that("hybrid weights sum to 1 for different p and degree 2", {
   deg <- 2L
-  expect_error(input_sampling(4L, deg = deg))
-  expect_error(input_sampling(5L, deg = deg))
+  expect_error(input_sampling(4L, deg = deg, feature_names = LETTERS[1:p]))
+  expect_error(input_sampling(5L, deg = deg, feature_names = LETTERS[1:p]))
   
   for (p in 6:20) {
-    pa <- input_partly_exact(p, deg = deg)
-    sa <- input_sampling(p, m = 100L, deg = deg, paired = FALSE)
+    pa <- input_partly_exact(p, deg = deg, feature_names = LETTERS[1:p])
+    sa <- input_sampling(
+      p, m = 100L, deg = deg, paired = FALSE, feature_names = LETTERS[1:p]
+    )
     expect_equal(sum(pa$w) + sum(sa$w), 1L)
   }
 })
 
 test_that("partly_exact_Z(p, k) fails for bad p or k", {
-  expect_error(partly_exact_Z(0L, k = 1L))
-  expect_error(partly_exact_Z(5L, k = 3L))
-  expect_error(partly_exact_Z(5L, k = 0L))
+  expect_error(partly_exact_Z(0L, k = 1L, feature_names = LETTERS[1:p]))
+  expect_error(partly_exact_Z(5L, k = 3L, feature_names = LETTERS[1:p]))
+  expect_error(partly_exact_Z(5L, k = 0L, feature_names = LETTERS[1:p]))
 })
 
 test_that("input_partly_exact(p, deg) fails for bad p or deg", {
-  expect_error(input_partly_exact(2L, deg = 0L))
-  expect_error(input_partly_exact(5L, deg = 3L))
+  expect_error(input_partly_exact(2L, deg = 0L, feature_names = LETTERS[1:p]))
+  expect_error(input_partly_exact(5L, deg = 3L, feature_names = LETTERS[1:p]))
 })
-
 
 # Helper functions
 test_that("head_list(x) = head(x) for matrix x", {
@@ -125,17 +136,7 @@ test_that("head_list(x)[[1L]] = head(x[[1L]]) for list of matries x", {
 })
 
 test_that("regoranize_list() fails for non-list inputs", {
-  expect_error(reorganize_list(alist = 1:10, nms = NULL))
-})
-
-test_that("weighted_colMeans() fails for non-matrix inputs", {
-  expect_error(weighted_colMeans(1))
-})
-
-test_that("weighted_colMeans() fails if weights are not long enough", {
-  X <- cbind(1:3, 2:4)
-  expect_error(weighted_colMeans(X, w = 1))
-  expect_error(weighted_colMeans(X, w = 1:2))
+  expect_error(reorganize_list(alist = 1:10))
 })
 
 test_that("weighted_colMeans() gives the same as colMeans() in unweighted case", {
@@ -144,4 +145,3 @@ test_that("weighted_colMeans() gives the same as colMeans() in unweighted case",
   expect_equal(c(weighted_colMeans(X, w = c(1, 1, 1))), colMeans(X))
   expect_equal(c(weighted_colMeans(X, w = c(2, 2, 2))), colMeans(X))
 })
-